+++ tools/src/main/kotlin/csense/android/tools/fps/FpsWatcher.kt
@@ -11,7 +11,6 @@ import com.commonsense.android.kotlin.base.time.*
 import com.commonsense.android.kotlin.system.extensions.*
 import com.commonsense.android.kotlin.system.logging.*
 import kotlinx.coroutines.*
-import kotlinx.coroutines.android.*
 import java.lang.ref.*
 import kotlin.math.*